+++ driver-core/src/test/java/com/datastax/driver/core/querybuilder/QueryBuilderTest.java
@@ -599,4 +599,10 @@ public class QueryBuilderTest {
             // Ok, that's what we expect
         }
     }
+
+    @Test(groups = "unit")
+    public void truncateTest() throws Exception {
+        assertEquals(truncate("foo").toString(), "TRUNCATE foo;");
+        assertEquals(truncate("foo", quote("Bar")).toString(), "TRUNCATE foo.\"Bar\";");
+    }
 }